# feat: support `client["/endpoint"].GET()` style calls