--- a/.changeset/stale-donuts-smoke.md
+++ b/.changeset/stale-donuts-smoke.md
@@ -1,0 +1,5 @@
+---
+"openapi-fetch": minor
+---
+
+Add support for `client["/endpoint"].GET()` style calls