--- a/packages/openapi-fetch/src/index.js
+++ b/packages/openapi-fetch/src/index.js
@@ -176,39 +176,15 @@ export default function createClient(clientOptions) {
     return { error, response };
   }
 
-  return {
-    /** Call a GET endpoint */
-    async GET(url, init) {
-      return coreFetch(url, { ...init, method: "GET" });
-    },
-    /** Call a PUT endpoint */
-    async PUT(url, init) {
-      return coreFetch(url, { ...init, method: "PUT" });
-    },
-    /** Call a POST endpoint */
-    async POST(url, init) {
-      return coreFetch(url, { ...init, method: "POST" });
-    },
-    /** Call a DELETE endpoint */
-    async DELETE(url, init) {
-      return coreFetch(url, { ...init, method: "DELETE" });
-    },
-    /** Call a OPTIONS endpoint */
-    async OPTIONS(url, init) {
-      return coreFetch(url, { ...init, method: "OPTIONS" });
-    },
-    /** Call a HEAD endpoint */
-    async HEAD(url, init) {
-      return coreFetch(url, { ...init, method: "HEAD" });
-    },
-    /** Call a PATCH endpoint */
-    async PATCH(url, init) {
-      return coreFetch(url, { ...init, method: "PATCH" });
-    },
-    /** Call a TRACE endpoint */
-    async TRACE(url, init) {
-      return coreFetch(url, { ...init, method: "TRACE" });
-    },
+  const methods = ["GET", "PUT", "POST", "DELETE", "OPTIONS", "HEAD", "PATCH", "TRACE"];
+
+  const methodMembers = Object.fromEntries(
+    methods.map((method) => [method, (url, init) => coreFetch(url, { ...init, method })]),
+  );
+
+  const coreClient = {
+    ...methodMembers,
+
     /** Register middleware */
     use(...middleware) {
       for (const m of middleware) {
@@ -231,6 +207,19 @@ export default function createClient(clientOptions) {
       }
     },
   };
+
+  const handler = {
+    get: (coreClient, property) => {
+      if (property in coreClient) {
+        return coreClient[property];
+      }
+
+      // Assume the property is an URL.
+      return Object.fromEntries(methods.map((method) => [method, (init) => coreFetch(property, { ...init, method })]));
+    },
+  };
+
+  return new Proxy(coreClient, handler);
 }
 
 // utils