--- a/packages/openapi-fetch/test/index.test.ts
+++ b/packages/openapi-fetch/test/index.test.ts
@@ -4,6 +4,7 @@ import createClient, {
   type Middleware,
   type MiddlewareCallbackParams,
   type QuerySerializerOptions,
+  createPathBasedClient,
 } from "../src/index.js";
 import { server, baseUrl, useMockRequestHandler, toAbsoluteURL } from "./fixtures/mock-server.js";
 import type { paths } from "./fixtures/api.js";
@@ -21,6 +22,19 @@ afterEach(() => server.resetHandlers());
 afterAll(() => server.close());
 
 describe("client", () => {
+  it("generates all proper functions", () => {
+    const client = createClient<paths>();
+
+    expect(client).toHaveProperty("GET");
+    expect(client).toHaveProperty("PUT");
+    expect(client).toHaveProperty("POST");
+    expect(client).toHaveProperty("DELETE");
+    expect(client).toHaveProperty("OPTIONS");
+    expect(client).toHaveProperty("HEAD");
+    expect(client).toHaveProperty("PATCH");
+    expect(client).toHaveProperty("TRACE");
+  });
+
   describe("TypeScript checks", () => {
     it("marks data or error as undefined, but never both", async () => {
       const client = createClient<paths>({
@@ -1858,9 +1872,10 @@ describe("client", () => {
     });
   });
 
-  describe("URL as property style call", () => {
+  describe("path based client", () => {
     it("performs a call without params", async () => {
-      const client = createClient<paths>({ baseUrl });
+      const client = createPathBasedClient<paths>({ baseUrl });
+
       const { getRequest } = useMockRequestHandler({
         baseUrl,
         method: "get",
@@ -1871,13 +1886,23 @@ describe("client", () => {
     });
 
     it("performs a call with params", async () => {
-      const client = createClient<paths>({ baseUrl });
+      const client = createPathBasedClient<paths>({ baseUrl });
       const { getRequestUrl } = useMockRequestHandler({
         baseUrl,
         method: "get",
         path: "/blogposts/:post_id",
         status: 200,
-        body: { message: "OK" },
+        body: { title: "Blog post title" },
+      });
+
+      // Wrong method
+      // @ts-expect-error
+      await client["/blogposts/{post_id}"].POST({
+        params: {
+          // Unknown property `path`.
+          // @ts-expect-error
+          path: { post_id: "1234" },
+        },
       });
 
       await client["/blogposts/{post_id}"].GET({
@@ -1886,10 +1911,74 @@ describe("client", () => {
         params: { path: { post_id: 1234 } },
       });
 
-      await client["/blogposts/{post_id}"].GET({
+      const { data, error } = await client["/blogposts/{post_id}"].GET({
         params: { path: { post_id: "1234" } },
       });
+
       expect(getRequestUrl().pathname).toBe("/blogposts/1234");
+
+      // Check typing of data.
+      if (error) {
+        // Fail, but we need the if above for type inference.
+        expect(error).toBeUndefined();
+      } else {
+        // @ts-expect-error
+        data.not_a_blogpost_property;
+        // Check typing of result value.
+        expect(data.title).toBe("Blog post title");
+      }
+    });
+
+    it("performs a POST call", async () => {
+      const client = createPathBasedClient<paths>({ baseUrl });
+      const { getRequest } = useMockRequestHandler({
+        baseUrl,
+        method: "post",
+        path: "/anyMethod",
+      });
+      await client["/anyMethod"].POST();
+      expect(getRequest().method).toBe("POST");
+    });
+
+    it("performs a PUT call with a request body", async () => {
+      const mockData = { status: "success" };
+
+      const client = createPathBasedClient<paths>({ baseUrl });
+      const { getRequestUrl } = useMockRequestHandler({
+        baseUrl,
+        method: "put",
+        path: "/blogposts",
+        status: 201,
+        body: mockData,
+      });
+
+      await client["/blogposts"].PUT({
+        body: {
+          title: "New Post",
+          body: "<p>Best post yet</p>",
+          // Should be a number, not a Date.
+          // @ts-expect-error
+          publish_date: new Date("2023-03-31T12:00:00Z"),
+        },
+      });
+
+      const { data, error, response } = await client["/blogposts"].PUT({
+        body: {
+          title: "New Post",
+          body: "<p>Best post yet</p>",
+          publish_date: new Date("2023-03-31T12:00:00Z").getTime(),
+        },
+      });
+
+      // assert correct URL was called
+      expect(getRequestUrl().pathname).toBe("/blogposts");
+
+      // assert correct data was returned
+      expect(data).toEqual(mockData);
+      expect(response.status).toBe(201);
+
+      // assert error is empty
+      expect(error).toBeUndefined();
     });
   });
 });