## Changes

Allow calls of the form `client["/path"].GET(...)` by using a JavaScript proxy object.

## How to Review

Suggested order:
- index.js
- index.d.ts
- Everything else.

Points to watch out for:
- Performance
- Reliance on additional ECMAScript features (Proxy).
- Sufficient unit test coverage (IMO light but sufficient given the implementation).
- Exporting / non-exporting of type helpers.

## Checklist

- [x] Unit tests updated
- [x] `docs/` updated (if necessary)
- [x] `pnpm run update:examples` run (only applicable for openapi-typescript)