@drwpow 2024-08-01
Thanks for submitting this! But I don’t see an attached issue discussing why this is needed / what problem this is solving. I’m seeing a significant performance regression by using Proxies and additional logic just for what seems to be a minor API change? Could you talk more why this direction was sought? What prior art it references? Does it solve any problems for you?

I’m not necessarily against adding this, but as stated in the CONTRIBUTING guidelines, changing the API in a PR without any prior discussion isn’t recommended, especially if it impacts performance for all users like this PR does (sidenote: for many libraries, yes, things like Proxies are negligible; while we don’t want to micro-optimize to no end, this library being a fetch client strives to shave off every nanosecond possible, and this introduces the biggest slowdown, performance-wise, in this library to-date so we’d better weigh the cost of what we get in return!).