@gzm0 2024-08-01
:facepalm: sorry, my bad. This fixes (and is discussed in) #1526.

The primary use case is better type inference / go to definition. Another reason why I'm interested in this is thinner mocking (e.g. with jest-mock-extended).

```ts
import { paths } from "...";

const mockClient = mock<Client<paths>>();

mockClient["/path"].GET(...).resolvedValue(...);
```

If you have discouraging performance numbers we can investigate a pay-only-if-you-use it alternative (probably wrapping the result of `createClient`).

(please excuse my brevity, I'm in-between things, but that missing cross-ref was embarassing).