@drwpow 2024-08-01
> 🤦 sorry, my bad. This fixes (and is discussed in) https://github.com/openapi-ts/openapi-typescript/issues/1526.

Ah, thanks! That issue came to mind but just wanted to check that this built off that convo. No worries—I think there are some great arguments in there.

> If you have discouraging performance numbers we can investigate a pay-only-if-you-use it alternative (probably wrapping the result of createClient).

I’d love to explore this, yeah! Just edited my original comment with more info. Benchmarking in general is really tricky (it’s _very_ hard to automate regression tests for it, without investing significant time and resources into maintaining it), but in general I’d like to explore all possible options of avoiding using a Proxy for _everything_ because of the performance hit. It [slows down everything by quite a lot](https://romgrk.com/posts/optimizing-javascript#4-avoid-indirection) (see naïve implementation; I may have made a mistake but hopefully this gives some reference): 

![CleanShot 2024-08-01 at 10 06 52@2x](https://github.com/user-attachments/assets/b102ebb9-76a8-4786-8dcd-a14cfa0be642)

If this API just took a separate codepath entirely I think that would be ideal.