@gzm0 2024-08-01
TY for the quick feedback. I'll attempt a separate factory alternative tomorrow.

RE performance:
- If the JIT falls back to interpretation, that's definitely an issue.
- TBC we're on the same page: performance is more important than code size (at least in the KB range?). (Although with a separate `createClient` function it will not matter if a decent bundler is used downstream).