@drwpow 2024-08-01
> TBC we're on the same page: performance is more important than code size (at least in the KB range?). (Although with a separate createClient function it will not matter if a decent bundler is used downstream).

Yes that’s correct! In the README and docs we do mention the code size as a _proxy_ for performance, but at the end of the day performance > code size. Because this library is meant to be a typed `fetch()` replacement for their OpenAPI schema, the goal is that users should be able to swap `fetch()` calls for this without any drop in performance (while that’s technically unattainable—any code introduced has some runtime—we want to at least make the performance hit so negligible it’s not a concern)