@gzm0 2024-08-02
It does not look like the proxy has any measurable performance impact:

Without the memoization:
```
 ✓ test/index.bench.js (16) 11283ms
   ✓ setup (4) 3562ms
     name                                hz     min     max    mean     p75     p99    p995    p999      rme  samples
   · openapi-fetch               963,767.47  0.0008  5.9080  0.0010  0.0009  0.0021  0.0024  0.0290   ±2.39%   481884   fastest
   · openapi-fetch (path based)  958,197.98  0.0009  0.5356  0.0010  0.0010  0.0020  0.0023  0.0029   ±0.44%   479099
   · openapi-typescript-fetch    442,678.98  0.0011  5.8343  0.0023  0.0013  0.0048  0.0055  0.0176  ±10.71%   221340
   · axios                       102,057.14  0.0089  0.5606  0.0098  0.0094  0.0169  0.0192  0.0335   ±0.50%    51029   slowest
   ✓ get (only URL) (6) 3861ms
     name                              hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · openapi-fetch               1,374.44  0.5659  5.6875  0.7276  0.7245  4.4769  5.1658  5.6875  ±4.72%      688
   · openapi-fetch (path based)  1,521.29  0.4680  8.1287  0.6573  0.6306  3.8165  5.9574  8.1287  ±6.24%      761
   · openapi-typescript-fetch    1,616.65  0.4747  6.3194  0.6186  0.5989  1.1297  5.1087  6.3194  ±5.51%      809   fastest
   · openapi-typescript-codegen  1,550.53  0.4916  5.5255  0.6449  0.6238  4.7036  5.0065  5.5255  ±5.04%      776
   · axios                         423.41  1.8487  6.3936  2.3618  2.4666  6.1766  6.2922  6.3936  ±4.03%      212   slowest
   · superagent                    503.85  1.6337  5.9643  1.9847  1.9958  5.3915  5.6255  5.9643  ±3.51%      253
   ✓ get (headers) (6) 3855ms
     name                              hz     min      max    mean     p75     p99    p995     p999      rme  samples
   · openapi-fetch               1,681.39  0.4554   5.8572  0.5947  0.5841  4.6528  4.8856   5.8572   ±5.40%      841
   · openapi-fetch (path based)  1,479.07  0.4598  37.2277  0.6761  0.5962  3.9853  5.4106  37.2277  ±15.19%      740
   · openapi-typescript-fetch    1,709.66  0.4375   7.1737  0.5849  0.5561  4.6950  4.9280   7.1737   ±5.73%      855   fastest
   · openapi-typescript-codegen  1,679.25  0.4550   5.4615  0.5955  0.5749  4.6242  4.8575   5.4615   ±5.27%      840
   · axios                         461.21  1.7771   6.6207  2.1682  2.1944  5.0415  5.9881   6.6207   ±3.53%      231   slowest
   · superagent                    480.56  1.5588   5.8003  2.0809  2.1562  5.3616  5.7713   5.8003   ±3.42%      241


 BENCH  Summary

  openapi-fetch - test/index.bench.js > setup
    1.01x faster than openapi-fetch (path based)
    2.18x faster than openapi-typescript-fetch
    9.44x faster than axios

  openapi-typescript-fetch - test/index.bench.js > get (only URL)
    1.04x faster than openapi-typescript-codegen
    1.06x faster than openapi-fetch (path based)
    1.18x faster than openapi-fetch
    3.21x faster than superagent
    3.82x faster than axios

  openapi-typescript-fetch - test/index.bench.js > get (headers)
    1.02x faster than openapi-fetch
    1.02x faster than openapi-typescript-codegen
    1.16x faster than openapi-fetch (path based)
    3.56x faster than superagent
    3.71x faster than axios
```

With the memoization:

```
 ✓ test/index.bench.js (16) 11160ms
   ✓ setup (4) 3438ms
     name                                hz     min      max    mean     p75     p99    p995    p999      rme  samples
   · openapi-fetch               898,597.93  0.0009   2.0690  0.0011  0.0010  0.0021  0.0025  0.0265   ±1.00%   449299   fastest
   · openapi-fetch (path based)  329,820.97  0.0013  10.6270  0.0030  0.0016  0.0045  0.0058  0.0147  ±12.57%   166556
   · openapi-typescript-fetch    425,604.10  0.0011   6.3861  0.0023  0.0013  0.0046  0.0053  0.0309  ±11.77%   212803
   · axios                       111,898.57  0.0083   0.8467  0.0089  0.0088  0.0129  0.0165  0.0228   ±0.50%    55950   slowest
   ✓ get (only URL) (6) 3828ms
     name                              hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · openapi-fetch               1,239.24  0.5681  7.3555  0.8069  0.7880  4.1139  5.2457  7.3555  ±5.21%      620
   · openapi-fetch (path based)  1,550.97  0.4764  5.5317  0.6448  0.6180  4.8220  5.1287  5.5317  ±5.08%      776
   · openapi-typescript-fetch    1,637.86  0.4533  5.6228  0.6106  0.5893  1.0828  5.0218  5.6228  ±5.00%      819   fastest
   · openapi-typescript-codegen  1,604.55  0.4969  5.6166  0.6232  0.6060  2.1447  5.0693  5.6166  ±5.09%      803
   · axios                         357.33  2.0225  6.9643  2.7986  2.9122  6.2065  6.9643  6.9643  ±3.21%      179   slowest
   · superagent                    402.58  1.8439  6.7624  2.4840  2.5649  6.3109  6.6520  6.7624  ±3.72%      202
   ✓ get (headers) (6) 3888ms
     name                              hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · openapi-fetch               1,501.18  0.4718  7.3496  0.6661  0.6251  4.9542  5.3748  7.3496  ±5.75%      755
   · openapi-fetch (path based)  1,736.68  0.4355  5.5869  0.5758  0.5641  2.3819  4.8996  5.5869  ±5.00%      869
   · openapi-typescript-fetch    1,783.00  0.4338  6.3428  0.5609  0.5448  4.4160  4.8103  6.3428  ±5.27%      892   fastest
   · openapi-typescript-codegen  1,694.24  0.4466  5.6741  0.5902  0.5613  4.3757  4.9665  5.6741  ±5.31%      848
   · axios                         375.33  1.9158  6.8901  2.6643  2.8660  6.4472  6.8901  6.8901  ±3.88%      188   slowest
   · superagent                    375.35  1.7544  6.9554  2.6642  2.7840  6.4310  6.9554  6.9554  ±3.90%      188


 BENCH  Summary

  openapi-fetch - test/index.bench.js > setup
    2.11x faster than openapi-typescript-fetch
    2.72x faster than openapi-fetch (path based)
    8.03x faster than axios

  openapi-typescript-fetch - test/index.bench.js > get (only URL)
    1.02x faster than openapi-typescript-codegen
    1.06x faster than openapi-fetch (path based)
    1.32x faster than openapi-fetch
    4.07x faster than superagent
    4.58x faster than axios

  openapi-typescript-fetch - test/index.bench.js > get (headers)
    1.03x faster than openapi-fetch (path based)
    1.05x faster than openapi-typescript-codegen
    1.19x faster than openapi-fetch
    4.75x faster than superagent
    4.75x faster than axios
```